add hints

+++ app/src/main/java/com/example/scrambledwordgame/data/Scores.kt
@@ -5,4 +5,4 @@ object Scores {
     val scores = mutableListOf<Score>()
 }
 
-data class Score(val name:String = "", val score:Int = 0, val time:Long)
+data class Score(val name:String = "", val score:Float = 0.0F, val time:Long)
